configure: allow specifying Python 3 sitepackages dir
OE's python patches to enable multilib and a lib directory that is not
"lib" do not cause distutils.sysconfig to return the correct libdir when
prefix is supplied. The 'sys.lib' variable added by OE is hard-coded to
the libdir at compile time, which would be the native Python's config
rather than cross.

Just pass it in from the recipe for now until this is gets fixed
upstream.